refactor: remove unused CSS files and update import paths
- Deleted unused CSS files for Header, Layout, Sidebar, and HelpModal components.
- Updated import paths in Layout and Sidebar components to reflect the new styles directory.
