--- apps/cardano-dashboard/src/components/Layout.module.css
@@ -1,27 +1,0 @@
-.layout {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: var(--background-dark);
-}
-
-.content {
-    display: flex;
-    flex: 1;
-    position: relative;
-}
-
-.main {
-    flex: 1;
-    padding: 2rem;
-    overflow-y: auto;
-    min-height: calc(100vh - 64px);
-    /* Subtract header height */
-}
-
-/* Mobile styles */
-@media (max-width: 768px) {
-    .main {
-        padding: 1rem;
-    }
-}
