+++ apps/cardano-dashboard/src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Sidebar from "./Sidebar";
 import Header from "./Header/Header";
-import styles from "./Layout.module.css";
+import styles from "../styles/Layout.module.css";
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
     return (
