--- apps/cardano-dashboard/src/components/Sidebar.module.css
@@ -1,52 +1,0 @@
-.sidebar {
-    width: 250px;
-    background-color: #1a1a1a;
-    color: #fff;
-    padding: 16px;
-    transition: width 0.3s ease;
-    position: relative;
-}
-
-.sidebar.collapsed {
-    width: 60px;
-    padding: 16px 8px;
-}
-
-.toggleButton {
-    position: absolute;
-    right: -12px;
-    top: 20px;
-    background-color: #1a1a1a;
-    border: none;
-    color: #fff;
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    padding: 0;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-}
-
-.toggleButton:hover {
-    background-color: #333;
-}
-
-.navItem {
-    display: flex;
-    align-items: center;
-    padding: 8px 0;
-    cursor: pointer;
-    white-space: nowrap;
-    overflow: hidden;
-}
-
-.navItem span {
-    margin-left: 8px;
-}
-
-.navItem:hover {
-    background-color: #333;
-}
